requirements.txt
Django==3.2.9
djangorestframework==3.12.4
gunicorn==20.0.4
django-crontab==0.7.1
django-cors-headers==3.7.0
djangorestframework-simplejwt==5
PyJWT==2.0.0
mysqlclient==1.4.2
pyyaml==5.4.1
requests==2.22.0
colorlog==4.6.2
python-dateutil==2.8.1
elasticsearch==6.0.0
pygsheets==2.0.3.1
redis==3.5.3
django-redis
moment
django-jsonfield-backport
websockets==6.0
pyee==6.0.0
pyppeteer